chore: refactor file structure + update eslint
