--- .editorconfig
@@ -1,14 +1,0 @@
-# http://editorconfig.org
-root = true
-
-[*]
-charset = utf-8
-end_of_line = crlf
-indent_size = 4
-indent_style = tab
-max_line_length = 160
-trim_trailing_whitespace = true
-
-[*.md]
-max_line_length = 0
-trim_trailing_whitespace = false
