+++ .github/ISSUE_TEMPLATE.md
@@ -1,11 +1,9 @@
-* **I'm submitting a ...**
-[ ] bug report
-[ ] feature request
-[ ] question about the decisions made in the repository
-[ ] question about how to use this project
+- **I'm submitting a ...**
+  [ ] bug report
+  [ ] feature request
+  [ ] question about the decisions made in the repository
+  [ ] question about how to use this project
 
-* **Summary**
+- **Summary**
 
-
-
-* **Other information** (e.g. detailed explanation, stacktraces, related issues, suggestions how to fix, links for us to have context, eg. StackOverflow, personal fork, etc.)
+- **Other information** (e.g. detailed explanation, stacktraces, related issues, suggestions how to fix, links for us to have context, eg. StackOverflow, personal fork, etc.)
