+++ .github/auto_assign.yml
@@ -5,12 +5,12 @@ addReviewers: true
 addAssignees: true
 
 # A list of reviewers to be added to pull requests (GitHub user name)
-reviewers: 
-  - ChewbaccaCookie
-  
-# A list of keywords to be skipped the process that add reviewers if pull requests include it 
+reviewers:
+    - ChewbaccaCookie
+
+# A list of keywords to be skipped the process that add reviewers if pull requests include it
 skipKeywords:
-  - wip
+    - wip
 
 # A number of reviewers added to the pull request
 # Set 0 to add all the reviewers (default: 0)
