--- .npmignore
@@ -1,14 +1,0 @@
-src
-test
-tsconfig.json
-tsconfig.module.json
-tslint.json
-.travis.yml
-.github
-.prettierignore
-.vscode
-build/docs
-**/*.spec.*
-coverage
-.nyc_output
-*.log
