+++ .vscode/settings.json
@@ -6,16 +6,7 @@
 		"**/.hg": true,
 		"**/CVS": true,
 		"**/.DS_Store": true,
-		"LICENSE": true,
-		"tsconfig.json": true,
-		".npmignore": true,
-		".prettierrc": true,
 		"dist": true,
-		".editorconfig": true,
-		"voicemeeter-connector.code-workspace": true,
-		".vscode": true,
-		"node_modules": true,
-		"docs": true,
-		".gitignore": true
+		"docs": true
 	}
 }
