+++ README.md
@@ -8,7 +8,7 @@ Voicemeeter Connector is a Node.js (Typescript) connector to use the official Vo
 | --------------- | ------------ |
 | <= 0.62         | 8,10         |
 | 1.x             | > 10, < 18   |
-| >= 2.0          | >= 18         |
+| >= 2.0          | >= 18        |
 
 > Node.js now includes build tools for Windows. You probably no longer need this tool. See https://github.com/felixrieseberg/windows-build-tools for details
