--- example/package.json
@@ -1,16 +1,0 @@
-{
-	"name": "voicemeeter-connector-example",
-	"version": "1.0.0",
-	"description": "",
-	"main": "javascript_example.js",
-	"scripts": {
-		"test-js": "node javascript_example.js",
-		"test-mjs": "node javascript_example.mjs",
-		"test-ts": "npx ts-node typescript_example.ts"
-	},
-	"dependencies": {
-		"voicemeeter-connector": "file:.."
-	},
-	"author": "",
-	"license": "ISC"
-}
