+++ examples/javascript-commonjs/package.json
@@ -1,0 +1,13 @@
+{
+	"name": "javascript-commonjs-example",
+	"private": true,
+	"type": "commonjs",
+	"description": "Example how to use the Voicemeeter Connector library in CommonJS project.",
+	"main": "example.js",
+	"scripts": {
+		"start": "node example.js"
+	},
+	"dependencies": {
+		"voicemeeter-connector": "*"
+	}
+}
