+++ lib/.npmignore
@@ -1,0 +1,4 @@
+node_modules
+src
+tsconfig.json
+.npmignore
