+++ lib/src/index.ts
@@ -1,0 +1,8 @@
+import * as constants from "./lib/constants";
+
+const { BusProperties, InterfaceTypes, StripProperties, MacroButtonModes } = constants;
+
+export { BusProperties, InterfaceTypes, MacroButtonModes, StripProperties };
+
+export { default as Voicemeeter } from "./lib/voicemeeter-connector";
+export * as types from "./types/voicemeeter-types";
