+++ package.json
@@ -1,42 +1,22 @@
 {
-	"name": "voicemeeter-connector",
-	"version": "2.1.0",
-	"description": "A Connector to use the Voicemeeter API",
-	"repository": "https://github.com/ChewbaccaCookie/voicemeeter-connector",
-	"license": "MIT",
-	"source": "src/index.ts",
-	"main": "dist/index.js",
-	"module": "dist/index.module.js",
-	"types": "dist/index.d.ts",
-	"scripts": {
-		"build": "microbundle",
-		"dev": "microbundle watch",
-		"lint": "npx eslint src/**/*.{js,ts,tsx} --quiet --fix",
-		"format": "npx prettier --write \"src/**/*.{ts,tsx,js,jsx}\"",
-		"generate-docu": "typedoc --out ./docs src/index.ts"
-	},
-	"dependencies": {
-		"koffi": "^2.10.1",
-		"winreg": "1.2.4"
-	},
-	"devDependencies": {
-		"@types/node": "17.0.21",
-		"@types/winreg": "1.2.31",
-		"@typescript-eslint/eslint-plugin": "5.13.0",
-		"@typescript-eslint/parser": "5.13.0",
-		"eslint": "8.10.0",
-		"eslint-config-airbnb-typescript": "16.1.0",
-		"eslint-config-prettier": "8.5.0",
-		"eslint-plugin-import": "2.25.4",
-		"eslint-plugin-prettier": "4.0.0",
-		"microbundle": "0.14.2",
-		"prettier": "2.5.1",
-		"typedoc": "0.22.14",
-		"typescript": "4.6.2"
-	},
-	"keywords": [
-		"voicemeeter",
-		"api",
-		"connector"
-	]
+    "name": "root",
+    "private": true,
+    "type": "module",
+    "description": "Root directory for the Voicemeeter Connector library. If you want to use the library, please install the package from the lib directory.",
+    "repository": "https://github.com/ChewbaccaCookie/voicemeeter-connector",
+    "workspaces": [
+        "lib",
+        "examples/*"
+    ],
+    "scripts": {
+        "format": "npx prettier --write .",
+        "lint": "tsc --project tsconfig.json --noEmit && eslint ."
+    },
+    "devDependencies": {
+        "@chewbaccacookie/eslint-config": "^1.1.1",
+        "@chewbaccacookie/prettier-config": "^1.2.0",
+        "eslint": "^9.29.0",
+        "prettier": "^3.5.3",
+        "typescript": "^5.8.3"
+    }
 }
