--- src/index.ts
@@ -1,7 +1,0 @@
-import Voicemeeter from "./lib/VoicemeeterConnector";
-import * as constants from "./lib/VoicemeeterConsts";
-import * as types from "./types/VoicemeeterTypes";
-
-const { BusProperties, InterfaceTypes, StripProperties, MacroButtonModes } = constants;
-
-export { Voicemeeter, BusProperties, InterfaceTypes, StripProperties, MacroButtonModes, types };
